# docs/swagger.yaml
basePath: /
definitions:
...
  gameConfig.ClusterIni:
    properties:
      bind_ip:
        type: string
      cluster_description:
        type: string
      cluster_intention:
        type: string
      cluster_key:
        type: string
      cluster_language:
        type: string
      cluster_name:
        type: string
      cluster_password:
        type: string
      console_enabled:
        description: '[MISC]'
        type: boolean
      game_mode:
        description: '[GAEMPLAY]'
        type: string
      lan_only_cluster:
        description: '[NETWORK]'
        type: boolean
      master_ip:
        type: string
      master_port:
        type: integer
      max_players:
        type: integer
      max_snapshots:
        type: integer
      offline_cluster:
        type: boolean
      pause_when_nobody:
        type: boolean
      pvp:
        type: boolean
      shard_enabled:
        description: '[SHARD]'
        type: boolean
      steam_group_admins:
        type: boolean
      steam_group_id:
        description: '[STEAM]'
        type: string
      steam_group_only:
        type: boolean
      tick_rate:
        type: integer
      vote_enabled:
        type: boolean
      vote_kick_enabled:
        type: boolean
      whitelist_slots:
        type: integer
    type: object
  gameConfig.ClusterIniConfig:
    properties:
      cluster:
        $ref: '#/definitions/gameConfig.ClusterIni'
      token:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        description: 提示代码
        type: integer
      data:
        description: 数据
      msg:
        description: 提示信息
        type: string
    type: object
host: localhost:8082
info:
  contact:
    name: API Support
    url: https://github.com/carrot-hu23/dst-admin-go
  description: 饥荒联机版服务器管理后台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DST Admin Go API
  version: "1.7.0"
paths:
...
  /api/game/config/adminlist:
    get:
      consumes:
      - application/json
      description: 获取房间 adminlist.txt 配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      summary: 获取房间 adminlist.txt 配置
      tags:
      - gameConfig
    post:
      consumes:
      - application/json
      description: 保存房间 adminlist.txt 配置
      parameters:
      - description: adminlist.txt 配置
        in: body
        name: list
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 保存房间 adminlist.txt 配置
      tags:
      - gameConfig
  /api/game/config/blacklist:
    get:
      consumes:
      - application/json
      description: 获取房间 blacklist.txt 配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      summary: 获取房间 blacklist.txt 配置
      tags:
      - gameConfig
    post:
      consumes:
      - application/json
      description: 保存房间 blacklist.txt 配置
      parameters:
      - description: blacklist.txt 配置
        in: body
        name: list
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 保存房间 blacklist.txt 配置
      tags:
      - gameConfig
  /api/game/config/clusterIni:
    get:
      consumes:
      - application/json
      description: 获取房间 cluster.ini 配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/gameConfig.ClusterIniConfig'
              type: object
      summary: 获取房间 cluster.ini 配置
      tags:
      - gameConfig
    post:
      consumes:
      - application/json
      description: 保存房间 cluster.ini 配置
      parameters:
      - description: cluster.ini 配置
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/gameConfig.ClusterIniConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 保存房间 cluster.ini 配置
      tags:
      - gameConfig
  /api/game/config/whitelist:
    get:
      consumes:
      - application/json
      description: 获取房间 whitelist.txt 配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      summary: 获取房间 whitelist.txt 配置
      tags:
      - gameConfig
    post:
      consumes:
      - application/json
      description: 保存房间 whitelist.txt 配置
      parameters:
      - description: whitelist.txt 配置
        in: body
        name: list
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 保存房间 whitelist.txt 配置
      tags:
      - gameConfig
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"